backend/requirements.txt
# ========================================
# CTRG Grant System - Python Dependencies
# ========================================
# Compatible with Python 3.9+

Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
django-environ>=0.11,<1.0
celery>=5.3,<6.0
redis>=5.0,<6.0
reportlab>=4.0,<5.0
Pillow>=10.0,<12.0
